+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/ChannelMessage.kt
@@ -15,7 +15,7 @@ data class ChannelMessage<T: MessageContent>(
     override val content: T,
     override val date: DateTime,
     override val editDate: DateTime?,
-    override val forwarded: ForwardedMessage?,
+    override val forwardInfo: ForwardInfo?,
     override val replyTo: Message?,
     override val replyMarkup: InlineKeyboardMarkup?,
     val authorSignature: AuthorSignature?