+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/AbleToBeForwardedMessage.kt
@@ -1,7 +1,0 @@
-package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
-
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.ForwardedMessage
-
-interface AbleToBeForwardedMessage : Message {
-    val forwarded: ForwardedMessage?
-}