+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/CommonMessage.kt
@@ -3,8 +3,8 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
 
 interface CommonMessage<T: MessageContent> : Message,
-    AbleToBeForwardedMessage,
-    AbleToBeEditedMessage,
-    AbleToReplyMessage,
-    AbleToBeMarkedUp,
+    PossiblyForwardedMessage,
+    PossiblyEditedMessage,
+    PossiblyReplyMessage,
+    PossiblyMarkedUp,
     ContentMessage<T>