+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/PossiblyMarkedUp.kt
@@ -2,6 +2,6 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
 
-interface AbleToBeMarkedUp {
+interface PossiblyMarkedUp {
     val replyMarkup: InlineKeyboardMarkup?
 }