+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/PossiblyPaymentMessage.kt
@@ -2,6 +2,6 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.payments.abstracts.PaymentInfo
 
-interface AbleToBePaymentedMessage : Message {
+interface PossiblyPaymentMessage : Message {
     val paymentInfo: PaymentInfo?
 }